Grid created with js
Used a bit of math and variables in order to make it so that the grid will adapt to the size of the screen when you reload the page. Unfortunately it only works with you reload the page, so I might try to fix that later. Also it uses a nested for loop in order to create the grid shape.